+++ app/actions/abstract/modal.jsx
@@ -1,0 +1,13 @@
+export default class ModalActions {
+  show() {
+    return true;
+  }
+
+  hide() {
+    return true;
+  }
+
+  reset() {
+    return true;
+  }
+}