+++ app/components/application/layout.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import HeaderLayout from 'components/header/layout';
 import TodoModal from 'components/todo_modal/view';
 import SigninModal from 'components/signin_modal/view';
+import SignupModal from 'components/signup_modal/view';
 import Footer from 'components/footer/view';
 import styles from './styles';
 
@@ -14,6 +15,7 @@ export default class ApplicationLayout extends React.Component {
           { this.props.children }
           <TodoModal/>
           <SigninModal/>
+          <SignupModal/>
         </main>
         <Footer/>
       </div>