+++ app/stores/abstract/modal.jsx
@@ -1,0 +1,15 @@
+export default class ModalStore {
+  constructor() {
+    this.showModal = false;
+  }
+
+  show() {
+    this.showModal = true;
+  }
+
+  hide() {
+    this.showModal = false;
+  }
+
+  reset() {}
+}