+++ app/stores/todo.jsx
@@ -1,11 +1,13 @@
 import Alt from 'alt_flux';
 import { createStore } from 'alt-utils/lib/decorators';
 import TodoActions from 'actions/todo';
+import ModalStore from 'stores/abstract/modal';
 
 @createStore(Alt)
-export default class TodosStore {
+export default class TodosStore extends ModalStore {
   constructor() {
-    this.showModal = false;
+    super();
+
     this.todo = {
       name: '',
       isComplete: false
@@ -19,19 +21,11 @@ export default class TodosStore {
     });
   }
 
-  show() {
-    this.showModal = true;
-  }
-
-  hide() {
-    this.showModal = false;
+  reset() {
+    this.todo.name = '';
   }
 
   setName(name) {
     this.todo.name = name;
-  }
-
-  reset() {
-    this.todo.name = '';
   }
 }